Place the plane with the landmarks
BUG: To place the plane we use the landmarks listed in the Markups
list. When clicking the button ‘+’ to add the landmarks, a
predetermined list named P1 is enabled to add the landmarks.

ENH: The plane actor and mapper is created only once. An observer is
added to check if the fiducial is moved. Since the objects are created
only once, the plane position is refreshed when moving a fiducial