Manage multiple planes
ENH: The angle plane widget is able to manage multiple planes.
Each plane is synchronized with a separate list in the fiducially
markups view.
The saving is done using the library pickle that uses python
dictionaries
The loading is done using pickle and load the dictionary that was saved
before.