Saving and loading
BUG The function to save and load the planes is connected to click
action. The problem is that there is a parameter set on the filename.
For this reason the dialog to save or load a file does not appear.
To correct the bug a new function is added called onSavePlanes that
calls the function savePlanes without the filename.